+++ front/app/containers/Admin/settings/customize/index.tsx
@@ -67,7 +67,7 @@ const ContrastWarning = styled(Warning)`
   margin-top: 10px;
 `;
 
-const WideSectionField = styled(SectionField)`
+export const WideSectionField = styled(SectionField)`
   max-width: calc(${(props) => props.theme.maxPageWidth}px - 100px);
 `;
 
@@ -385,6 +385,13 @@ class SettingsCustomizeTab extends PureComponent<
     }
   };
 
+  getSetting = (setting: string) => {
+    return (
+      get(this.state.attributesDiff, `settings.${setting}`) ??
+      get(this.state.tenant, `data.attributes.settings.${setting}`)
+    );
+  };
+
   handleColorPickerOnClick = () => {
     this.setState({ colorPickerOpened: true });
   };
@@ -439,7 +446,8 @@ class SettingsCustomizeTab extends PureComponent<
     const { tenant } = this.state;
     if (!tenant?.data.attributes.settings.events_page) return;
 
-    const previousValue = tenant.data.attributes.settings.events_page.enabled;
+    const previousValue = this.getSetting('events_page.enabled');
+
     this.setState((state) => {
       return {
         attributesDiff: {
@@ -450,6 +458,30 @@ class SettingsCustomizeTab extends PureComponent<
             events_page: {
               ...get(state.settings, 'events_page', {}),
               ...get(state.attributesDiff, 'settings.events_page', {}),
+              enabled: !previousValue,
+            },
+          },
+        },
+      };
+    });
+  };
+
+  handleToggleEventsWidget = () => {
+    const { tenant } = this.state;
+    if (!tenant?.data.attributes.settings.events_widget) return;
+
+    const previousValue = this.getSetting('events_widget.enabled');
+
+    this.setState((state) => {
+      return {
+        attributesDiff: {
+          ...state.attributesDiff,
+          settings: {
+            ...state.settings,
+            ...get(state.attributesDiff, 'settings', {}),
+            events_widget: {
+              ...get(state.settings, 'events_widget', {}),
+              ...get(state.attributesDiff, 'settings.events_widget', {}),
               enabled: !previousValue,
             },
           },
@@ -521,13 +553,7 @@ class SettingsCustomizeTab extends PureComponent<
                     </Label>
                     <ColorPickerInput
                       type="text"
-                      value={
-                        get(attributesDiff, `settings.core.${colorName}`) ||
-                        get(
-                          tenant,
-                          `data.attributes.settings.core.${colorName}`
-                        )
-                      }
+                      value={this.getSetting(`core.${colorName}`)}
                       onChange={this.handleColorPickerOnChange(colorName)}
                     />
                     {contrastRatioWarningOfColor && contrastRatioOfColor && (
@@ -696,39 +722,44 @@ class SettingsCustomizeTab extends PureComponent<
             </WideSectionField>
           </Section>
 
-          {tenant.data.attributes.settings?.events_page &&
-            tenant.data.attributes.settings.events_page.allowed && (
-              <Section>
-                <SectionTitle>
-                  <FormattedMessage {...messages.eventsSection} />
-                </SectionTitle>
+          {tenant.data.attributes.settings?.events_page?.allowed && (
+            <Section>
+              <SectionTitle>
+                <FormattedMessage {...messages.eventsSection} />
+              </SectionTitle>
 
-                <WideSectionField>
-                  <Setting>
-                    <ToggleLabel>
-                      <StyledToggle
-                        checked={
-                          get(attributesDiff, 'settings.events_page.enabled') ??
-                          get(
-                            tenant,
-                            'data.attributes.settings.events_page.enabled'
-                          )
-                        }
-                        onChange={this.handleToggleEventsPage}
-                      />
-                      <LabelContent>
-                        <LabelTitle>
-                          {formatMessage(messages.eventsPageSetting)}
-                        </LabelTitle>
-                        <LabelDescription>
-                          {formatMessage(messages.eventsPageSettingDescription)}
-                        </LabelDescription>
-                      </LabelContent>
-                    </ToggleLabel>
-                  </Setting>
-                </WideSectionField>
-              </Section>
-            )}
+              <WideSectionField>
+                <Setting>
+                  <ToggleLabel>
+                    <StyledToggle
+                      checked={this.getSetting('events_page.enabled')}
+                      onChange={this.handleToggleEventsPage}
+                    />
+                    <LabelContent>
+                      <LabelTitle>
+                        {formatMessage(messages.eventsPageSetting)}
+                      </LabelTitle>
+                      <LabelDescription>
+                        {formatMessage(messages.eventsPageSettingDescription)}
+                      </LabelDescription>
+                    </LabelContent>
+                  </ToggleLabel>
+                </Setting>
+              </WideSectionField>
+
+              {tenant.data.attributes.settings?.events_widget?.allowed && (
+                <Outlet
+                  id="app.containers.Admin.settings.customize.EventsWidgetSwitch"
+                  checked={this.getSetting('events_widget.enabled')}
+                  onChange={this.handleToggleEventsWidget}
+                  title={formatMessage(messages.eventsWidgetSetting)}
+                  description={formatMessage(
+                    messages.eventsWidgetSettingDescription
+                  )}
+                />
+              )}
+            </Section>
+          )}
 
           <SubmitWrapper
             loading={this.state.loading}