+++ front/app/modules/commercial/insights/admin/containers/Insights/Edit/InputsTable/ScanCategory.tsx
@@ -1,4 +1,4 @@
-import React, { useMemo } from 'react';
+import React, { useMemo, useEffect, useState } from 'react';
 import { withRouter, WithRouterProps } from 'react-router';
 
 // styles
@@ -13,7 +13,6 @@ import { isNilOrError } from 'utils/helperUtils';
 
 // components
 import Button from 'components/UI/Button';
-import { Spinner } from 'cl2-component-library';
 
 // services
 import { insightsTriggerCategoriesSuggestionsTasks } from 'modules/commercial/insights/services/insightsCategoriesSuggestionsTasks';
@@ -43,26 +42,30 @@ const ScanContainer = styled.div`
   }
 `;
 
-const ScanButtonContent = styled.div`
-  display: flex;
-  align-items: center;
-  > div:first-child {
-    margin-right: 8px;
-  }
-`;
-
 const ScanCategory = ({
   intl: { formatMessage },
   params: { viewId },
   location: { query },
 }: InjectedIntlProps & WithRouterProps) => {
+  const [loading, setLoading] = useState(false);
   const categories = useMemo(() => [query.category], [query.category]);
   const categorySuggestionsPendingTasks = useInsightsCategoriesSuggestionsTasks(
-    viewId
+    viewId,
+    { categories }
   );
+
+  useEffect(() => {
+    if (
+      !isNilOrError(categorySuggestionsPendingTasks) &&
+      categorySuggestionsPendingTasks.length === 0
+    ) {
+      setLoading(false);
+    }
+  }, [categorySuggestionsPendingTasks]);
 
   const suggestCategories = async () => {
     try {
+      setLoading(true);
       await insightsTriggerCategoriesSuggestionsTasks(viewId, categories);
     } catch {
       // Do nothing
@@ -72,14 +75,6 @@ const ScanCategory = ({
   if (isNilOrError(categorySuggestionsPendingTasks)) {
     return null;
   }
-
-  const loading = Boolean(
-    categorySuggestionsPendingTasks.find((pendingTask) =>
-      pendingTask.relationships?.categories.data
-        .map((category) => category.id)
-        .includes(query.category)
-    )
-  );
 
   return (
     <ScanContainer data-testid="insightsScanCategory">
@@ -96,11 +91,9 @@ const ScanCategory = ({
         buttonStyle="admin-dark"
         onClick={suggestCategories}
         disabled={loading}
+        processing={loading}
       >
-        <ScanButtonContent>
-          {loading && <Spinner size="22px" />}
-          {formatMessage(messages.categoriesEmptyScanButton)}
-        </ScanButtonContent>
+        {formatMessage(messages.categoriesEmptyScanButton)}
       </Button>
     </ScanContainer>
   );